+++ forge-gui-desktop/src/main/java/forge/screens/match/controllers/CField.java
@@ -79,7 +79,7 @@ public class CField implements ICDoc {
                     final Input ipm = controller.getInputQueue().getInput();
                     if (ipm instanceof InputPayMana && ipm.getOwner().equals(player)) {
                         final int oldMana = player.getMana(colorCode);
-                        controller.useMana(colorCode.byteValue());
+                        controller.useMana(colorCode);
                         return oldMana != player.getMana(colorCode);
                     }
                 }
